refacor tables, add shadcn
