+++ app/api/standings/utils/calculateCustomStandings.ts
@@ -32,11 +32,11 @@ export function calculateCustomStandings(
 
     eventMatches.forEach((match) => {
       if (match.league_entry_1_points > match.league_entry_2_points) {
-        playerMap[match.league_entry_1].win_points += 1;
+        playerMap[match.league_entry_1].win_points += 3;
         playerMap[match.league_entry_1].total_wins += 1;
         playerMap[match.league_entry_2].total_losses += 1;
       } else if (match.league_entry_1_points < match.league_entry_2_points) {
-        playerMap[match.league_entry_2].win_points += 1;
+        playerMap[match.league_entry_2].win_points += 3;
         playerMap[match.league_entry_2].total_wins += 1;
         playerMap[match.league_entry_1].total_losses += 1;
       } else {
